add command input field logic and computer mesh

+++ Assets/Scripts/Commander/Map/HorizontalMapScale.cs
@@ -14,7 +14,7 @@ public class HorizontalMapScale : MapScale
 
     protected override void SetScaleLinePos(float delta, float bigLineSteps)
     {
-        float start = transform.position.x - delta - bigLineSteps * maxScaleGaps / 2;
+        float start = transform.position.x - delta - bigLineSteps * (int)(maxScaleGaps / 2);
         float steps = bigLineSteps * 0.5f;
 
         for (int i = 0; i < mapScaleLines.Length; ++i)
